+++ src/mod.ts
@@ -1,6 +1,6 @@
 // export { LimetteApp } from "./server/app.ts";
-export { LimetteApp } from "./router/app.ts";
+export { LimetteApp } from "./server/app.ts";
 export { dev } from "./dev/dev.ts";
 export type { AppRootOptions } from "./server/ssr.ts";
 export type { Handlers } from "./server/router.ts";
-export type { Context } from "./router/app.ts";
+export type { Context } from "./server/app.ts";
