+++ src/types.ts
@@ -1,7 +1,7 @@
 // @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
 import type { LitElement } from "lit";
 import type { Handlers } from "./server/router.ts";
-import type { MiddlewareFn } from "./router/middleware.ts";
+import type { MiddlewareFn } from "./server/middleware.ts";
 
 export interface RouteConfig {
   skipInheritedLayouts: boolean; // Skip already inherited layouts
